Fixed steal skin

+++ src/main/java/net/wurstclient/altmanager/screens/AltEditorScreen.java
@@ -9,10 +9,20 @@ package net.wurstclient.altmanager.screens;
 
 import java.io.IOException;
 import java.io.InputStream;
+import java.net.MalformedURLException;
 import java.net.URI;
+import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 
+import com.google.gson.Gson;
+import com.google.gson.JsonArray;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonObject;
+import com.google.gson.JsonParser;
+
+import org.apache.commons.codec.binary.Base64;
+import org.apache.commons.io.IOUtils;
 import org.lwjgl.glfw.GLFW;
 import org.lwjgl.opengl.GL11;
 
@@ -153,24 +163,92 @@ public abstract class AltEditorScreen extends Screen
 	
 	private final String stealSkin(String name)
 	{
-		String skin = name + ".png";
-		
-		URI u = URI.create("http://skins.minecraft.net/MinecraftSkins/")
-			.resolve(skin);
-		Path path = skinFolder.resolve(skin);
-		
+
 		createSkinFolder();
+		Path path = skinFolder.resolve(name+".png");
 		
-		try(InputStream in = u.toURL().openStream())
-		{
-			Files.copy(in, path);
-			return "\u00a7a\u00a7lSaved skin as " + skin;
+		if(path.toFile().exists()) { //Skin already exists, no need to download skin again.
+			return "\u00a7a\u00a7lSaved skin as " + name + ".png";
+		}
+		
+		try{
+
+			URI u = getSkinUrl(name);
 			
+			InputStream in = u.toURL().openStream();
+			
+			// TODO finish this
+			Files.copy(in, path);
+			return "\u00a7a\u00a7lSaved skin as " + name + ".png";
 		}catch(IOException e)
 		{
 			e.printStackTrace();
 			return "\u00a74\u00a7lSkin could not be saved.";
+		}catch(NullPointerException iu) {
+			iu.printStackTrace();
+			return "\u00a74\u00a7lPlayer does not exist.";
 		}
+	}
+	
+	public URI getSkinUrl(String username) throws MalformedURLException, IOException {
+		/*
+		 * Will get a url using the username.
+		 * 
+		 * First it needs to grabe the UUID of the username. It will be able to do this by going to 
+		 * this url https://api.mojang.com/users/profiles/minecraft/<username>
+		 * 
+		 * When you get the uuid then you have to go to 
+		 * https://sessionserver.mojang.com/session/minecraft/profile/<UUID> and grab the value of textures witch is
+		 * a base64 encoded text of another JSON queary
+		 * 
+		 * looking something like this
+		 * 
+		 * {
+		 *	  "timestamp" : number,
+		 *	  "profileId" : "uuid",
+		 *	  "profileName" : "username",
+		 *	  "textures" : {
+		 *	    "SKIN" : {
+		 *	      "url" : "url to skin",
+		 *	      "metadata" : {
+		 *	        "model" : "slim"
+		 *	      }
+		 *	    }
+		 *	  }
+		 *	}
+		 * 
+		 * And we return the url 
+		 * 
+		 */
+		
+		//Getting players uuid
+		URI UUID_url = URI.create("https://api.mojang.com/users/profiles/minecraft/").resolve(username);
+		InputStream UUIDinputStream = UUID_url.toURL().openStream();
+		
+		JsonObject uuidJsonObj = new Gson().fromJson(IOUtils.toString(UUIDinputStream, StandardCharsets.UTF_8), JsonObject.class);
+		String uuid = uuidJsonObj.get("id").getAsString();
+		
+		
+		//Getting base64 test with uuid
+		URI sessionURL = URI.create("https://sessionserver.mojang.com/session/minecraft/profile/").resolve(uuid);
+		InputStream sessionInputStream = sessionURL.toURL().openStream();
+		JsonObject sessionJsonObj = new Gson().fromJson(IOUtils.toString(sessionInputStream, StandardCharsets.UTF_8), JsonObject.class);
+		JsonArray propJArray = sessionJsonObj.get("properties").getAsJsonArray();
+		JsonObject zeroJObj = propJArray.get(0).getAsJsonObject();
+		String jB64 = zeroJObj.get("value").toString();	
+		
+		//Decode to string
+		byte[] valueDecoded = Base64.decodeBase64(jB64.getBytes());
+		
+		//Grab url for skin
+		JsonObject bDecode = new Gson().fromJson(new String(valueDecoded), JsonObject.class);
+		JsonObject tJObj = bDecode.get("textures").getAsJsonObject();
+		JsonObject skinJObj = tJObj.get("SKIN").getAsJsonObject();
+		String skin = skinJObj.get("url").getAsString();
+		
+		//converts the string into a url
+		URI skinUrl = URI.create(skin);
+		return skinUrl;
 	}
 	
 	@Override
